some suggestions and bugfixes for existing bugs
