+++ backend/src/Designer/Views/Preview/index.cshtml
@@ -13,10 +13,6 @@
       height: 100vh;
       width: 100vw;
       }
-      div {
-      height: 100%;
-      width: 100%;
-      }
     </style>
     </head>
     <body>
