+++ frontend/app-preview/src/PreviewApp.module.css
@@ -1,6 +1,14 @@
 .previewContainer {
-  height: 100%;
-  width: auto;
+  --toolbar-height: 80px;
+  --subtoolbar-height: 60px;
+  --header-border-bottom-width: 1px;
+  --header-height: calc(
+    var(--toolbar-height) + var(--subtoolbar-height) + var(--header-border-bottom-width)
+  );
+  --left-menu-width: 68px;
+
+  min-height: 100vh;
+  width: 100%;
   display: flex;
   flex-direction: column;
 }
