--- frontend/app-preview/src/components/AppPreviewHeader/index.tsx
@@ -1,1 +1,0 @@
-export { AppPreviewHeader } from './AppPreviewHeader';
