+++ frontend/app-preview/src/components/AppPreviewSubMenu.module.css
@@ -1,0 +1,35 @@
+.leftContent svg {
+  color: #fff;
+}
+
+.rightContent {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.subHeader {
+  color: white;
+  border-bottom: 1px solid gray;
+  background-color: #0c6536;
+  display: flex;
+  align-items: center;
+  height: var(--subtoolbar-height);
+  margin-top: 80px;
+}
+
+.viewSizeButtons {
+  margin: 12px 15px 15px 0;
+  --component-toggle_button-color-border-inactive: white;
+  --component-toggle_button-color-border-active: white;
+  --component-toggle_button-color-background-active: white;
+  --component-toggle_button-color-background-inactive: transparent;
+  --component-toggle_button-color-text-active: black;
+  --component-toggle_button-color-text-inactive: white;
+}
+
+/* Overriding this style because it is hardcoded in the design system */
+.viewSizeButtons button[aria-pressed='false']:hover {
+  background-color: #dbf5d5 !important;
+  color: black !important;
+}
