+++ frontend/packages/shared/src/components/altinnHeader/AltinnHeader.module.css
@@ -14,9 +14,6 @@
 .preview {
   background-color: #118849;
   color: #fff;
-  width: 100%;
-  position: fixed;
-  padding: 16px;
 }
 
 .altinnHeaderBar .leftContent {
